Move RFI test to  their own file
